+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Updated docs for deprecated methods `_update_queue_mode` and `get_queue_mode` in `Project`
     * Use the `queue_mode` attribute in `Project` to get and set the queue mode instead
     * For more information, visit https://docs.labelbox.com/reference/migrating-to-workflows#upcoming-changes
+* Updated `project.export_labels` to support filtering by start/end time formats "YYYY-MM-DD" and "YYYY-MM-DD hh:mm:ss"
 ### Fixed
 
 # Version 3.27.1 (2022-09-16)